♻️ refactor(new.tsx): extract event creation and input handling logic into hooks
The commit reorganizes the code by:
1. Moving event creation logic to useCreateEvent hook
2. Extracting input initialization to useInitializeInput hook
3. Creating useKeyboardHeight hook for keyboard animations
4. Improving code maintainability and reusability
