+++ apps/expo/src/hooks/useInitializeInput.ts
@@ -1,0 +1,100 @@
+import { useCallback, useEffect, useState } from "react";
+
+import type { RecentPhoto } from "~/store";
+import { useAppStore } from "~/store";
+
+const VALID_IMAGE_REGEX = /^[\w.:\-_/]+\|\d+(\.\d+)?\|\d+(\.\d+)?$/;
+
+interface UseInitializeInputParams {
+  text?: string;
+  imageUri?: string;
+  recentPhotos: RecentPhoto[];
+}
+
+export function useInitializeInput({
+  text,
+  imageUri,
+  recentPhotos,
+}: UseInitializeInputParams) {
+  const [initialized, setInitialized] = useState(false);
+  const {
+    setInput,
+    setImagePreview,
+    setLinkPreview,
+    setActiveInput,
+    setIsOptionSelected,
+  } = useAppStore();
+
+  const handleImagePreview = useCallback(
+    (uri: string) => {
+      setImagePreview(uri);
+      setInput(uri.split("/").pop() || "");
+    },
+    [setImagePreview, setInput],
+  );
+
+  const handleLinkPreview = useCallback(
+    (url: string) => {
+      setLinkPreview(url);
+      setInput(url);
+    },
+    [setLinkPreview, setInput],
+  );
+
+  useEffect(() => {
+    setInput("");
+    setImagePreview(null);
+    setLinkPreview(null);
+
+    if (text) {
+      setInput(text);
+      setActiveInput("describe");
+      setIsOptionSelected(true);
+    } else if (imageUri) {
+      if (VALID_IMAGE_REGEX.test(imageUri)) {
+        const [uri, width, height] = imageUri.split("|");
+        if (uri) {
+          if (uri.startsWith("http")) {
+            handleLinkPreview(uri);
+            setActiveInput("url");
+          } else {
+            void handleImagePreview(uri);
+            setActiveInput("upload");
+          }
+        }
+        setInput(`Image: ${width ?? "unknown"}x${height ?? "unknown"}`);
+        setIsOptionSelected(true);
+      } else {
+        console.warn("Invalid image URI format:", imageUri);
+        setActiveInput("describe");
+        setIsOptionSelected(false);
+      }
+    } else {
+      const mostRecentPhoto = recentPhotos[0];
+      console.log("Initializing with most recent photo:", mostRecentPhoto);
+      if (mostRecentPhoto?.uri) {
+        setActiveInput("upload");
+        setIsOptionSelected(true);
+        void handleImagePreview(mostRecentPhoto.uri);
+      } else {
+        setActiveInput("describe");
+        setIsOptionSelected(false);
+      }
+    }
+
+    setInitialized(true);
+  }, [
+    text,
+    imageUri,
+    handleImagePreview,
+    handleLinkPreview,
+    recentPhotos,
+    setActiveInput,
+    setImagePreview,
+    setInput,
+    setIsOptionSelected,
+    setLinkPreview,
+  ]);
+
+  return { initialized };
+}
